refactor(log): move getSampleIntervalMinutes to log.util.ts
Standalone utility functions belong in *.util.ts files per the repo
convention. Keep the function's behaviour and JSDoc unchanged; only
the location and imports move.
